New code: Add code to set up HPO process.
This commit seeks to provide code to set up a hyperparameter tuning job. We have several hyperparameters that we seek to optimize, such as batch size, learning rate, and epoch count. To determine an ideal combination in our search space, we will train several models as part of HPO and pick the model HP's with the highest validation accuracy.

Addresses section 2 of Project 3.