Feat: Added finalized hpo.py file
In order to train our HPO job and find ideal hyperparameters, we will
need an entry point in the form of a Python script. This commit includes
a finalized version that has been run and tested on AWS Sagemaker.

The file is called hpo.py to refer to the job it is performing.