Docs: Add README.md to project
A README.md file is important for those who visit this repo, as it is among
the first pages that one spots. This document contains some instructions to
set up the project in AWS, as well as some background information about the
ML project. Additionally, this documents answers some questions about the
approach we took to train and deploy our model and reflects on the results.